+++ src/main/scala/com/bundlepricing/repos/BundleRepoComponent.scala
@@ -7,19 +7,24 @@ import org.bson.types.ObjectId
 import com.bundlepricing.domains.Bundle
 
 trait BundleRepoComponent {
+  
   def bundleRepo: BundleRepo
 
-  abstract class BundleRepo(val dbName: String = "", val collectionName: String = "")
-                           (implicit val context: Context,
-                             val mongoClient: MongoClient,
-                             val idManifest: Manifest[ObjectId],
-                             val entityManifest: Manifest[Bundle])
-                 extends RepoMetaData with Repository {
-    type Id = ObjectId
+  trait BundleRepo extends RepoMetaData with Repository {
     type Key = String
     type Entity = Bundle
-    
+
     val keyField = "key"
     def key(entity: Bundle): Key = entity.key
   }
+  
+  abstract class BundleMongoRepo(val dbName: String = "", val collectionName: String = "")
+                                (implicit val context: Context, val mongoClient: MongoClient)
+                 extends SalatRepoMataData with BundleRepo {
+    type Id = ObjectId
+    
+    val idManifest = implicitly[Manifest[Id]]
+    val entityManifest = implicitly[Manifest[Entity]]
+  }
+  
 }
