+++ src/main/scala/com/bundlepricing/repos/ItemRepoComponent.scala
@@ -7,19 +7,24 @@ import org.bson.types.ObjectId
 import com.bundlepricing.domains.Item
 
 trait ItemRepoComponent {
+  
   def itemRepo: ItemRepo
 
-  abstract class ItemRepo(val dbName: String = "", val collectionName: String = "")
-                         (implicit val context: Context,
-                             val mongoClient: MongoClient,
-                             val idManifest: Manifest[ObjectId],
-                             val entityManifest: Manifest[Item])
-                  extends RepoMetaData with Repository {
-    type Id = ObjectId
+  trait ItemRepo extends RepoMetaData with Repository {
     type Key = String
     type Entity = Item
-    
+                             
     val keyField = "name"
     def key(entity: Item): Key = entity.name
   }
+  
+  abstract class ItemMongoRepo(val dbName: String = "", val collectionName: String = "")
+                         (implicit val context: Context, val mongoClient: MongoClient)
+                  extends SalatRepoMataData with ItemRepo {
+    type Id = ObjectId
+    
+    val idManifest = implicitly[Manifest[Id]]
+    val entityManifest = implicitly[Manifest[Entity]]
+  }
+  
 }
